some minor tweaks: added a yamllint config file so it stops erroring on 'long' lines removed heimdahl from the setup as it's not being used since the short links to services is not possible.
I really need to make time to catch up on Nina stream. But tonight I need to write and do other stuff and Thursday is NSSH dang it ￼
￼
apache tweaks:
removing bad forwards for websocket connections to the wrong locations. ws: should only be required for the python noteobok
no other locations appear to require one at this time.

add redis handling to images for notebook, neuroglancer, and cloud volume launchers

re-enabled the notebook server

python notebook added with initial work towards code demos for interacting with the redis database.

neuroglancer launcher configured to store it's key, host, and port in the redis database for retrieval as a dictionary later.