Addition of a new notebook file to test access to the docker socket from inside the notebook container.
Docker compose changes to correctly launch the notebooks using the start-notebook script so that we can set uid/gid values on the way up;
- setting nb_uid/gid automatically on startup not yet handled.

cleaned up some junk options that were commented out in docker-compose, removed cruft notes at the bottom of the file that were no longer correct.

removed igneous from the cloudvolume dockerfile, this is no longer required

removed old dockerfile cruft for neuroglancer docker file based on attempting to use the node application directly instead of using the python wrapper

addition of pre-notebook hook script for scratchpad notebook container. This adds the docker group to the container on launch and puts the local user in the group.
- this allows notebooks to interact directly with the docker socket

added docker to pip install list for notebook

begun work on a rebuild command, this script will:
- clean up docker-compose cruft, dead containers, networks
- rebuild all containers (not using --no-cache)
- recreate the shared container network so it's a consistent named network
- create tagged 'latest' versions of the cloud volume and neuroglancer images so they can be invoked directly